+++ CHANGELOG.md
@@ -48,6 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for code:load_abs/1, code:load_binary/3
 - Added support for loading / closing AVMPacks at runtime
 - Added support for ESP-IDF v5.x
+- Added support for `calendar:system_time_to_universal_time/2`
 
 ### Fixed
 - Fixed issue with formatting integers with io:format() on STM32 platform